dotfiles: inherit environmental variables when X session is started
Inherit environment variables from .zshenv. This is to fix PATH
variables when launching from a login manager or an interactive login
what happens is that appended paths to PATH variable are not available
to programs like dwm which cause problems such as it not being able to
run local scripts in appended paths to PATH variable as this appending
was previously invoked only when a shell is opened.

[1] https://wiki.archlinux.org/index.php/LightDM#Optional_configuration_and_tweaks

Signed-off-by: Daniel Jacob Chittoor <djchittoor47@gmail.com>